Working framework without settings file

+++ nv_regression/core_tool/views.py
@@ -1,9 +1,14 @@
 from django.shortcuts import render
-from .models import processTracker
+from .models import processTracker, vbios, systems
 import subprocess, os
 from rest_framework.response import Response
 from django.http import HttpResponse
 import datetime
+from .serializers import vbiosSerializer, processTrackerSerializer, systemsSerializer
+import psutil
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.views import APIView
 
 # Create your views here.
 class processTracker(APIView):
@@ -12,6 +17,28 @@ class processTracker(APIView):
         processDetails = processTracker.objects.all()
         serializer = processTrackerSerializer(processDetails, many=True)
         return Response(serializer.data)
+
+class vbiosInfo(APIView):
+
+    def get(self,request):
+        vbiosDetails = vbios.objects.all()
+        serializer = vbiosSerializer(vbiosDetails, many=True)
+        return Response(serializer.data)
+    def post(self,request):
+        serializer = vbiosSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class systemsDetail(APIView):
+
+    def get(self,request):
+        systemDetails = systems.objects.all()
+        serializer = systemsSerializer(systemDetails, many=True)
+        return Response(serializer.data)
+
 
 def startTest(request):
     testDetails = {}
@@ -107,34 +134,46 @@ def startTest(request):
 
 
     #cwd
-    testDetails["cwd"] = request.POST["CWD"]
+    testDetails["cwd"] = request.POST["CWD"].strip()
 
     p = subprocess.popen("python test.py")
     pid = p.pid
 
     saveProc(pid)
 
-    return redirect('/currProc')
+    return redirect('/links')
 
 
-def saveProc(request,pid):
-    processObject = processTracker(userEmail=request.user.email, procName=request.POST['pName'], procId=int(pid), timeCreated=datetime.datetime.now())
+def saveProc(request,procId):
+    processObject = processTracker(userEmail=request.user.email, procName=request.POST['pName'], procId=int(procId), timeCreated=datetime.datetime.now())
     processObject.save()
-    os.waitpid(int(pid))
-    killTest(int(pid))
+    os.waitpid(int(procId))
+    killTest(int(procId))
 
 def killTest(request,procId):
 
     if request.user.is_authenticated:
-        proc = processTracker.objects.get(procId=int(procId))
-        if proc.userEmail.strip() == request.user.email.strip():
-            os.kill(int(procId))
-            proc.delete()
+        try:
+            proc = processTracker.objects.get(procId=int(procId))
+            if proc.userEmail.strip() == request.user.email.strip():
+                proc.delete()
+                if psutil.pid_exists(int(procId)):
+                    os.kill(int(procId))
+        except:
+            return HttpResponse("This process does not exist")
+
         else:
             return HttpResponse("You are not authorised to kill this process")
     else:
         return HttpResponse("Login First")
 
 
+def testInit(request):
+    return render(request,'testInit.html')
 
+def home(request):
+    return render(request,'home.html')
+
+def killInit(request):
+    return render(request,'killTest.html')
